Added 'Allot_Funds' for Project Task Area and fixed 32/64 bit issue with modifier flags.

+++ src/Meta_Class.cpp
@@ -2438,7 +2438,7 @@ void Meta_Class::impl::impl_Generate( )
    if( get_obj( ).child_Modifier( ).iterate_forwards( modifier_key_info ) )
    {
       string all_modifiers;
-      unsigned flag_value( 0x100u );
+      uint64_t flag_value( 0x100u );
 
       do
       {
